+++ RELEASE_NOTES.md
@@ -1,3 +1,7 @@
+### 5.1.9
+- removed System.Reactive depdendency
+- raised FSharp.Core version to >= 4.7.0
+
 ### 5.1.8
 
 - added polygon transformataions (#57)